+++ src/formats/gvox_raw.cpp
@@ -26,19 +26,19 @@ GVoxRawContext::GVoxRawContext() = default;
 
 auto GVoxRawContext::create_payload(GVoxScene scene) -> GVoxPayload {
     GVoxPayload result = {};
-    result.size += sizeof(size_t);
-    for (size_t node_i = 0; node_i < scene.node_n; ++node_i) {
+    result.size += sizeof(uint64_t);
+    for (uint64_t node_i = 0; node_i < scene.node_n; ++node_i) {
         if (scene.nodes[node_i].voxels == nullptr) {
             continue;
         }
-        result.size += sizeof(size_t) * 3;
+        result.size += sizeof(uint64_t) * 3;
         result.size += scene.nodes[node_i].size_x * scene.nodes[node_i].size_y * scene.nodes[node_i].size_z * sizeof(GVoxVoxel);
     }
-    result.data = new uint8_t[result.size];
+    result.data = new uint8_t[static_cast<size_t>(result.size)];
     auto *buffer_ptr = (uint8_t *)result.data;
     std::memcpy(buffer_ptr, &scene.node_n, sizeof(scene.node_n));
     buffer_ptr += sizeof(scene.node_n);
-    for (size_t node_i = 0; node_i < scene.node_n; ++node_i) {
+    for (uint64_t node_i = 0; node_i < scene.node_n; ++node_i) {
         if (scene.nodes[node_i].voxels == nullptr) {
             continue;
         }
@@ -48,8 +48,8 @@ auto GVoxRawContext::create_payload(GVoxScene scene) -> GVoxPayload {
         buffer_ptr += sizeof(scene.nodes[node_i].size_y);
         std::memcpy(buffer_ptr, &scene.nodes[node_i].size_z, sizeof(scene.nodes[node_i].size_z));
         buffer_ptr += sizeof(scene.nodes[node_i].size_z);
-        size_t const voxels_size = scene.nodes[node_i].size_x * scene.nodes[node_i].size_y * scene.nodes[node_i].size_z * sizeof(GVoxVoxel);
-        std::memcpy(buffer_ptr, scene.nodes[node_i].voxels, voxels_size);
+        uint64_t const voxels_size = scene.nodes[node_i].size_x * scene.nodes[node_i].size_y * scene.nodes[node_i].size_z * sizeof(GVoxVoxel);
+        std::memcpy(buffer_ptr, scene.nodes[node_i].voxels, static_cast<size_t>(voxels_size));
         buffer_ptr += voxels_size;
     }
     return result;
@@ -63,20 +63,20 @@ auto GVoxRawContext::parse_payload(GVoxPayload payload) -> GVoxScene {
     GVoxScene result = {};
     auto *buffer_ptr = (uint8_t *)payload.data;
     uint8_t *buffer_sentinel = (uint8_t *)payload.data + payload.size;
-    result.node_n = *(size_t *)buffer_ptr;
+    result.node_n = *(uint64_t *)buffer_ptr;
     buffer_ptr += sizeof(result.node_n);
-    result.nodes = (GVoxSceneNode *)std::malloc(sizeof(GVoxSceneNode) * result.node_n);
-    size_t node_i = 0;
+    result.nodes = (GVoxSceneNode *)std::malloc(sizeof(GVoxSceneNode) * static_cast<size_t>(result.node_n));
+    uint64_t node_i = 0;
     while (buffer_ptr < buffer_sentinel) {
-        result.nodes[node_i].size_x = *(size_t *)buffer_ptr;
+        result.nodes[node_i].size_x = *(uint64_t *)buffer_ptr;
         buffer_ptr += sizeof(result.nodes[node_i].size_x);
-        result.nodes[node_i].size_y = *(size_t *)buffer_ptr;
+        result.nodes[node_i].size_y = *(uint64_t *)buffer_ptr;
         buffer_ptr += sizeof(result.nodes[node_i].size_y);
-        result.nodes[node_i].size_z = *(size_t *)buffer_ptr;
+        result.nodes[node_i].size_z = *(uint64_t *)buffer_ptr;
         buffer_ptr += sizeof(result.nodes[node_i].size_z);
-        size_t const voxels_size = result.nodes[node_i].size_x * result.nodes[node_i].size_y * result.nodes[node_i].size_z * sizeof(GVoxVoxel);
-        result.nodes[node_i].voxels = (GVoxVoxel *)std::malloc(voxels_size);
-        std::memcpy(result.nodes[node_i].voxels, buffer_ptr, voxels_size);
+        uint64_t const voxels_size = result.nodes[node_i].size_x * result.nodes[node_i].size_y * result.nodes[node_i].size_z * sizeof(GVoxVoxel);
+        result.nodes[node_i].voxels = (GVoxVoxel *)std::malloc(static_cast<size_t>(voxels_size));
+        std::memcpy(result.nodes[node_i].voxels, buffer_ptr, static_cast<size_t>(voxels_size));
         buffer_ptr += voxels_size;
         ++node_i;
     }
